PTQuantizerInsertionCommand
